# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PSKit Research Agent API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/dependencies": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Dependency Status",
        "description": "Report live dependency status to an authenticated administrator.",
        "operationId": "dependency_status_api_v1_admin_dependencies_get",
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DependencyReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/af3/jobs/{job_id}/gpu-usage": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Reconcile Af3 Gpu Usage",
        "description": "Correct one AF3 job's settled GPU usage with an audit reason.",
        "operationId": "reconcile_af3_gpu_usage_api_v1_admin_af3_jobs__job_id__gpu_usage_put",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GpuReconciliationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Af3Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/af3/jobs/{job_id}/gpu-usage/audit": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Af3 Gpu Reconciliations",
        "description": "List the audited GPU accounting changes for an AF3 job.",
        "operationId": "list_af3_gpu_reconciliations_api_v1_admin_af3_jobs__job_id__gpu_usage_audit_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Af3GpuReconciliation"
                  },
                  "title": "Response List Af3 Gpu Reconciliations Api V1 Admin Af3 Jobs  Job Id  Gpu Usage Audit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/skills/{skill_id}/versions/{version}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Register Skill Version",
        "description": "Register an immutable version of an administrator-managed Skill.",
        "operationId": "register_skill_version_api_v1_admin_skills__skill_id__versions__version__put",
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterSkillRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/limits": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Set User Limits",
        "description": "Set a user's monthly token and daily GPU quotas.",
        "operationId": "set_user_limits_api_v1_admin_users__user_id__limits_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLimitsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/skills": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Set User Skill Grants",
        "description": "Replace the Skill IDs granted to a user.",
        "operationId": "set_user_skill_grants_api_v1_admin_users__user_id__skills_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          },
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSkillGrantsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSkillGrantsRequest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Af3GpuReconciliation": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "source": {
            "type": "string",
            "enum": [
              "worker",
              "admin"
            ],
            "title": "Source"
          },
          "previous_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Minutes"
          },
          "actual_minutes": {
            "type": "integer",
            "title": "Actual Minutes"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "user_id",
          "source",
          "previous_minutes",
          "actual_minutes",
          "reason",
          "created_at"
        ],
        "title": "Af3GpuReconciliation"
      },
      "Af3Job": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "estimated_gpu_minutes": {
            "type": "integer",
            "title": "Estimated Gpu Minutes"
          },
          "actual_gpu_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Gpu Minutes"
          },
          "gpu_accounting_status": {
            "type": "string",
            "enum": [
              "reserved",
              "released",
              "settled",
              "pending_reconciliation",
              "reconciled"
            ],
            "title": "Gpu Accounting Status",
            "default": "reserved"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "artifacts": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "fold_input": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fold Input"
          },
          "simulation": {
            "type": "boolean",
            "title": "Simulation",
            "default": false
          },
          "resource_requirements": {
            "$ref": "#/components/schemas/ComputeResourceRequirements"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "progress",
          "estimated_gpu_minutes"
        ],
        "title": "Af3Job"
      },
...
      "CatalogItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CatalogItem"
      },
...
      "ComputeResourceRequirements": {
        "properties": {
          "capability": {
            "type": "string",
            "const": "af3",
            "title": "Capability",
            "default": "af3"
          },
          "gpu_count": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Gpu Count",
            "default": 1
          },
          "min_gpu_memory_mb": {
            "type": "integer",
            "maximum": 1048576.0,
            "minimum": 0.0,
            "title": "Min Gpu Memory Mb",
            "default": 0
          }
        },
        "type": "object",
        "title": "ComputeResourceRequirements"
      },
...
      "DependencyReport": {
        "properties": {
          "identity": {
            "type": "string",
            "enum": [
              "available",
              "unavailable",
              "unauthorized",
              "rate_limited",
              "unknown",
              "partial",
              "mock",
              "disabled",
              "inbound_only"
            ],
            "title": "Identity"
          },
          "model_gateway": {
            "type": "string",
            "enum": [
              "available",
              "unavailable",
              "unauthorized",
              "rate_limited",
              "unknown",
              "partial",
              "mock",
              "disabled",
              "inbound_only"
            ],
            "title": "Model Gateway"
          },
          "mcp": {
            "type": "string",
            "enum": [
              "available",
              "unavailable",
              "unauthorized",
              "rate_limited",
              "unknown",
              "partial",
              "mock",
              "disabled",
              "inbound_only"
            ],
            "title": "Mcp"
          },
          "af3": {
            "type": "string",
            "enum": [
              "available",
              "unavailable",
              "unauthorized",
              "rate_limited",
              "unknown",
              "partial",
              "mock",
              "disabled",
              "inbound_only"
            ],
            "title": "Af3"
          },
          "mcp_servers": {
            "additionalProperties": {
              "type": "string",
              "enum": [
                "available",
                "unavailable",
                "unauthorized",
                "rate_limited",
                "unknown",
                "partial",
                "mock",
                "disabled",
                "inbound_only"
              ]
            },
            "type": "object",
            "title": "Mcp Servers"
          }
        },
        "type": "object",
        "required": [
          "identity",
          "model_gateway",
          "mcp",
          "af3",
          "mcp_servers"
        ],
        "title": "DependencyReport"
      },
...
      "GpuQuota": {
        "properties": {
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "used": {
            "type": "integer",
            "title": "Used"
          },
          "reserved": {
            "type": "integer",
            "title": "Reserved"
          },
          "remaining": {
            "type": "integer",
            "title": "Remaining"
          },
          "unit": {
            "type": "string",
            "const": "gpu_minutes",
            "title": "Unit"
          },
          "period": {
            "type": "string",
            "const": "day",
            "title": "Period"
          },
          "resets_at": {
            "type": "string",
            "format": "date-time",
            "title": "Resets At"
          }
        },
        "type": "object",
        "required": [
          "limit",
          "used",
          "reserved",
          "remaining",
          "unit",
          "period",
          "resets_at"
        ],
        "title": "GpuQuota"
      },
      "GpuReconciliationRequest": {
        "properties": {
          "actual_gpu_minutes": {
            "type": "integer",
            "maximum": 1440.0,
            "minimum": 0.0,
            "title": "Actual Gpu Minutes"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 5,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "actual_gpu_minutes",
          "reason"
        ],
        "title": "GpuReconciliationRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuotaCounter": {
        "properties": {
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "used": {
            "type": "integer",
            "title": "Used"
          },
          "reserved": {
            "type": "integer",
            "title": "Reserved"
          },
          "remaining": {
            "type": "integer",
            "title": "Remaining"
          },
          "unit": {
            "type": "string",
            "const": "tokens",
            "title": "Unit"
          },
          "period": {
            "type": "string",
            "enum": [
              "day",
              "month"
            ],
            "title": "Period"
          },
          "resets_at": {
            "type": "string",
            "format": "date-time",
            "title": "Resets At"
          }
        },
        "type": "object",
        "required": [
          "limit",
          "used",
          "reserved",
          "remaining",
          "unit",
          "period",
          "resets_at"
        ],
        "title": "QuotaCounter"
      },
...
      "RegisterSkillRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Description"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Tools"
          },
          "instructions": {
            "type": "string",
            "maxLength": 65536,
            "minLength": 1,
            "title": "Instructions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "description",
          "tools",
          "instructions"
        ],
        "title": "RegisterSkillRequest"
      },
...
      "StorageQuota": {
        "properties": {
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "used": {
            "type": "integer",
            "title": "Used"
          },
          "remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining"
          },
          "unit": {
            "type": "string",
            "const": "bytes",
            "title": "Unit",
            "default": "bytes"
          }
        },
        "type": "object",
        "required": [
          "limit",
          "used",
          "remaining"
        ],
        "title": "StorageQuota"
      },
...
      "UsageSnapshot": {
        "properties": {
          "tokens": {
            "$ref": "#/components/schemas/QuotaCounter"
          },
          "gpu": {
            "$ref": "#/components/schemas/GpuQuota"
          },
          "storage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StorageQuota"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "tokens",
          "gpu"
        ],
        "title": "UsageSnapshot"
      },
...
      "UserLimitsRequest": {
        "properties": {
          "token_monthly_limit": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Token Monthly Limit"
          },
          "gpu_daily_minutes": {
            "type": "integer",
            "maximum": 1440.0,
            "minimum": 0.0,
            "title": "Gpu Daily Minutes"
          }
        },
        "type": "object",
        "required": [
          "token_monthly_limit",
          "gpu_daily_minutes"
        ],
        "title": "UserLimitsRequest"
      },
      "UserSkillGrantsRequest": {
        "properties": {
          "skill_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skill Ids"
          }
        },
        "type": "object",
        "required": [
          "skill_ids"
        ],
        "title": "UserSkillGrantsRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}